refactor: 💡 organized components and layouts and textbubble (#87)
