+++ src/components/common/Icons.jsx
@@ -1,0 +1,29 @@
+import { ReactComponent as ArrowLeft } from '../../assets/icon/arrow-left.svg';
+import { ReactComponent as SendMsg } from '../../assets/icon/send-msg.svg';
+import { ReactComponent as ArrowDown } from '../../assets/icon/arrow-down.svg';
+import { ReactComponent as HomeTitleS } from '../../assets/icon/home-title--S.svg';
+import { ReactComponent as HomeTitleC } from '../../assets/icon/home-title--c.svg';
+import { ReactComponent as HomeTitleR } from '../../assets/icon/home-title--r.svg';
+import { ReactComponent as HomeTitleU } from '../../assets/icon/home-title--u.svg';
+import { ReactComponent as HomeTitleM } from '../../assets/icon/home-title--m.svg';
+import { ReactComponent as HomeTitleQuestionMark } from '../../assets/icon/home-title--question-mark.svg';
+import { ReactComponent as HomeTitleExclamationMark } from '../../assets/icon/home-title--exclamation-mark.svg';
+import { ReactComponent as Shared } from '../../assets/icon/shared.svg';
+
+export const Icons = {
+  ArrowLeft,
+  SendMsg,
+  ArrowDown,
+  Shared,
+  HomeTitle: {
+    S: HomeTitleS,
+    C: HomeTitleC,
+    R: HomeTitleR,
+    U: HomeTitleU,
+    M: HomeTitleM,
+    QuestionMark: HomeTitleQuestionMark,
+    ExclamationMark: HomeTitleExclamationMark,
+  },
+};
+
+export default Icons;
