+++ src/components/common/index.js
@@ -1,0 +1,4 @@
+export * from './BgBubble';
+export * from './Button';
+export * from './BlurBlockBg';
+export * from './Icons';
