+++ src/components/index.js
@@ -1,9 +1,4 @@
-export * from './BgBubble';
-export * from './Icons';
-export * from './Button';
-export * from './BlurBlockBg';
+export * from './common';
 export * from './chat';
 export * from './introduction';
-export * from './ChatRoom';
-export * from './DragLayout';
-export * from './introduction/IntroductionLayout';
+export * from './layout';
