+++ src/components/introduction/index.js
@@ -1,3 +1,4 @@
 export * from './SpeechBubble';
 export * from './Tag';
 export * from './ScrumMessageBox';
+export * from './Role';
