+++ src/components/layout/ChatRoom.jsx
@@ -9,8 +9,8 @@ import {
   ChatLog,
   ChoiceResponse,
   SubmitMsg,
-} from './chat';
-import { Button } from './Button';
+} from '../chat';
+import { Button } from '../common';
 
 export function ChatRoom({ props }) {
   const {
