+++ src/components/layout/index.js
@@ -1,0 +1,4 @@
+export * from './ChatRoom';
+export * from './IntroductionLayout';
+export * from './DragLayout';
+export * from './CoverLayout';
