+++ src/constants/homeInfo.js
@@ -18,7 +18,8 @@ export const HomeInfo = {
     {
       id: 3,
       content:
-        '馬上要參與Scrum專案了，<br />哪裡可以讓我快速了解？',
+        // eslint-disable-next-line react/jsx-filename-extension
+        ['馬上要參與Scrum專案了', <br />, '哪裡可以讓我快速了解？'],
     },
   ],
   button: '點我進入',
