+++ src/constants/thankInfo.js
@@ -18,7 +18,8 @@ export const ThankInfo = {
     {
       id: 3,
       content:
-        '馬上要參與Scrum專案了，<br />我相信我馬上就可以上手！',
+        // eslint-disable-next-line react/jsx-filename-extension
+        ['馬上要參與Scrum專案了', <br />, '我相信我馬上就可以上手！'],
     },
   ],
   bottomText: '我們希望能透過這個體驗讓您更加了解 Scrum 唷！ 謝謝您撥了點時間來體驗這整個 Scrum 的流程，希望對您們未來實務上能有所幫助！！',
