+++ src/views/exam/ExamProductBacklog.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/jsx-props-no-spreading */
 import { DragLayout } from '../../components';
 import { ExamProductBacklogInfo } from '../../constants/examProductBacklogInfo';
